🔨 Refactor(#28): axios 폴더구조 변경
