+++ src/axios/instance.ts
@@ -4,7 +4,7 @@ import {
   requestFunction,
   responseError,
   responseFunction,
-} from './intercepter';
+} from './interceptor';
 
 const api = axios.create({
   baseURL: import.meta.env.VITE_BASE_URL,
