+++ src/features/part/service/parsePartData.ts
