+++ src/features/section/apis.ts
@@ -1,4 +1,4 @@
-import api from '../../apis/axios/instance';
+import api from '../../axios/instance';
 import { Section } from './types';
 
 const sectionsApis = {
